implementation-python/requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
psycopg2-binary>=2.9.9

# LLM
litellm>=1.50.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0

duckduckgo-search>=7.0.0


